Unify Makefiles and change erisa-exec a bit
Modify Makefiles to unify CFLAGS and BUILD_DIR_ROOT in the
topmost file. Lower level Makefiles may append to those vars
as needed. Make output cleaner too. This is still not perfect but
its better than nothing and i am veyr commited to not using CMake
in this project.

Also change erisa-exec code a bit, since it was not properly initializing
VM memory with the ram array, also stop after every instruction and
wait for a kyestroke.