Implement erisa_asm() + minor changes
Add executable erisa-asm as well as the implementation of erisa_asm()
function in liberisa. It was necessary to add -fPIC to compile asm.c

Rename source files in liberisa to better reflect their function.

Add example erisa assembly in prototyping/ directory.

Assembler still needs a lot of cleanup, also its possible to codegen
the mapping structure. Error codes of erisa_asm() need to be specified
somewhere (there are many of them). It works for now though.